+++ core/src/overrides/env.ts
@@ -1,9 +1,9 @@
 // 重写环境变量
-import { OverridesProps } from "."
+import { OverridesProps } from "./../interface"
 
 export const restENV = (overrides: OverridesProps) => {
-  if(overrides.env){
-    Object.entries(overrides.env).forEach(([key,value])=>{
+  if (overrides.env) {
+    Object.entries(overrides.env).forEach(([key, value]) => {
       process.env[key] = value;
     })
   }
