+++ core/src/overrides/index.ts
@@ -1,14 +1,11 @@
 // 获取 根目录下 自己定义的配置
 
 import fs from 'fs';
-import { resolveModule, resolveApp, Paths,paths } from "./pathUtils"
+import { resolveModule, resolveApp, paths } from "./pathUtils"
 
-import webpack from "webpack"
 import { restENV } from "./env"
-import { overridePaths ,} from 'kkt/lib/overrides/paths';
-import { Application } from 'express';
-
-import { MockerOption } from "mocker-api"
+import { overridePaths, } from 'kkt/lib/overrides/paths';
+import { OverridesProps } from "./../interface"
 
 const tsOptions = {
   compilerOptions: {
@@ -29,57 +26,18 @@ const tsOptions = {
 
 const confPath = resolveModule(resolveApp, '.kktssrrc');
 
-export type Options = Omit<OverridesProps,"overridesClientWebpack"|"overridesServerWebpack"|"overridesCommonWebpack"|"overridesWebpack"> & {
-  env:"development" | "production"
-}
-
-export type WebpackConfigFunction = (conf: webpack.Configuration[] | webpack.Configuration, env: "development" | "production", options: Options) => webpack.Configuration[] | webpack.Configuration;
-
-export interface OverridesProps {
-  env?:{
-    /** 环境变变量 */
-    GENERATE_SOURCEMAP?: string,
-    INLINE_RUNTIME_CHUNK?: string,
-    ESLINT_NO_DEV_ERRORS?: string,
-    DISABLE_ESLINT_PLUGIN?: string,
-  },
-  /** paths 脚本中webpack配置 使用的地址  */
-  paths?: Partial<Paths>;
-
-  /** 最终覆写 webpack  配置 **/
-  /** 客户端配置  */
-  overridesClientWebpack?: (conf: webpack.Configuration, env: "development" | "production", options: Options) => webpack.Configuration,
-  /** 服务端配置  */
-  overridesServerWebpack?: (conf: webpack.Configuration, env: "development" | "production", options: Options) => webpack.Configuration;
-  /** 公共覆盖配置 */
-  overridesCommonWebpack?: (conf: webpack.Configuration, env: "development" | "production", options: Options) => webpack.Configuration;
-  // 最终的配置
-  overridesWebpack?: WebpackConfigFunction
-
-  /** 服务端打包入口 */
-  server_path?: string,
-  /** 客户端打包入口 */
-  client_path?: string,
-  /** 输出文件地址 */
-  output_path?: string;
-  /**  watch 配置 */
-  watchOptions?: webpack.Configuration["watchOptions"];
-  /** 代理 */ 
-  proxySetup?: (app: Application) => {
-    path: string | string[],
-    options?: MockerOption
-  }
-}
-
 let overrides: OverridesProps = {
-  env:{},
+  env: {},
   // 服务端打包入口
   server_path: resolveModule(resolveApp, 'src/server'),
   // 客户端打包入口
   client_path: resolveModule(resolveApp, 'src/client'),
   /** 输出文件地址 */
   output_path: resolveApp("dist"),
-
+  /** 是否使用原始 react-script 下的配置, 📢注意：这个不控制 server 配置， **/
+  isUseOriginalConfig: false,
+  /** 是否使用 server 配置 **/
+  isUseServerConfig: true,
   // paths 地址
   paths: {},
   // 自定义 client 配置设置
@@ -98,7 +56,7 @@ let overrides: OverridesProps = {
 
 export async function loaderConf(): Promise<OverridesProps> {
   let kktssrrc: OverridesProps = {};
-  
+
   try {
     if (fs.existsSync(confPath) && /.ts$/.test(confPath)) {
       require('ts-node').register(tsOptions);
