+++ core/src/utils/ProcessingConfig.ts
@@ -2,8 +2,7 @@
 import { reactScripts, webpackConfigPath } from "../overrides/pathUtils"
 import webpackNodeExternals from "webpack-node-externals"
 import webpack from "webpack"
-import { OptionsProps } from "../interface"
-import { OverridesProps } from "../overrides"
+import { OptionsProps, OverridesProps } from "../interface"
 import CreateTemporaryAsset from "./plugins/CreateTemporaryAsset"
 import { restOutPut } from "./output"
 import { getWbpackBarPlugins, restWebpackManifestPlugin, clearHtmlTemp, addServerPlugins, AddDefinePlugin } from "./plugins"
